- added support for ICU 57.1   - build tested with Python 2.7.5, Python 3.5.1 and PyPy 2.2.1   - release version 1.9.3

+++ common.cpp
@@ -59,7 +59,9 @@ U_STABLE void U_EXPORT2 _stopDecode(const void *context,
     if (chars && length)
     {
         const int size = stop->src_length - length + 1;
-        const int len = length < sizeof(stop->chars) - 1 ? length : sizeof(stop->chars) - 1;
+        const size_t len = (size_t) length < sizeof(stop->chars) - 1
+            ? (size_t) length
+            : sizeof(stop->chars) - 1;
 
         strncpy(stop->chars, chars, len); stop->chars[len] = '\0';
         stop->error_position = -1;
